Limit log prefix to only contain level/severity
Otherwise our logs contained the module name; since this software only consists
of a single module, this was redundant with journald's storage of the process
responsible for a given log element.

(It *is* possible that libraries we use will attempt to log through the Python
logging framework as well; however, that doesn't appear to be happening to any
substantial degree in practice).